Refactor webpack integration APIs
